BBBSL-26: Added checksum and fixed all the text cases
