+++ Gemfile
@@ -41,6 +41,7 @@ end
 
 group :test do
   gem 'fakeredis', github: 'guilleiguaran/fakeredis', ref: '2ebe19229954c7234bed019a0c5d28d5cf5b40f6'
+  gem 'minitest-stub_any_instance'
   gem 'webmock'
 end
