+++ test/test_helper.rb
@@ -5,6 +5,7 @@ require_relative '../config/environment'
 require 'rails/test_help'
 require 'fakeredis/minitest'
 require 'webmock/minitest'
+require 'minitest/stub_any_instance'
 
 module ActiveSupport
   class TestCase
